==> use batch openai/2. start_batch.ipynb <==
from openai import OpenAI

input_file_path = "./부대관리훈령_JSONL_RESULT.jsonl"

## 개인 API 키 입력
# client = OpenAI(api_key = "")
batch_input_file = client.files.create(
  file=open(input_file_path, "rb"),
  purpose="batch"
)

batch_input_file_id = batch_input_file.id

client.batches.create(
    input_file_id=batch_input_file_id,
    endpoint="/v1/chat/completions",
    completion_window="24h", #응답 요청 시간(240502기준 24h만 가능)
    metadata={
      "description": "국방부 훈령 1차"
    }
)

## 배치 상태 확인하기
client.batches.retrieve("").status

## 배치 취소하기
client.batches.cancel("")

## 아웃풋 확인
output_file_id = client.batches.retrieve("batch_5gwXWG9KF2bP8NGLL95ghZ0r").output_file_id
