==> use batch openai/4. concat_batch_result.ipynb <==
import json
import pandas as pd

file_path = './batch_국방부훈령_output.jsonl' ## 배치로 나온 ouptut 결과
raw_data_path = './국방부훈령_JSONL_RESULT.jsonl' ## 배치 전 데이터

## josnl 파일 불러오기
def load_jsonl(file_path):
    data = []
    with open(file_path, 'r', encoding='utf-8') as file:
        for line in file:
            data.append(json.loads(line.strip()))
    return data

## 배치 결과에서 답변만 뽑아내는 법
def extract_content(data):
    contents = []
    for item in data:
        if 'response' in item and 'body' in item['response'] and 'choices' in item['response']['body']:
            for choice in item['response']['body']['choices']:
                if 'message' in choice and 'content' in choice['message']:
                    contents.append(choice['message']['content'])
    return contents

## 배치 입력에서 user content만 뽑아내기
def extract_user_content(data):
    user_contents = []
    for item in data:
        if 'body' in item and 'messages' in item['body']:
            for message in item['body']['messages']:
                if message['role'] == 'user' and 'content' in message:
                    user_contents.append(message['content'])
    return user_contents

jsonl_data = load_jsonl(file_path)
raw_data = pd.read_excel('./국방부훈령_batch_input.xlsx')[['document','user']]
contents_data = extract_content(jsonl_data)
# user_prompt = extract_user_content(raw_data)

# data = pd.DataFrame({
#     'user_promt' : user_prompt,
#     'assistant' : contents_data
# })

raw_data['contents'] = contents_data
raw_data

# user_content를 "\n"을 기준으로 분리하고, 각 분리된 값을 새로운 행으로 추가
expanded_rows = []

for index, row in raw_data.iterrows():
    assistant = row['contents'].split('\n')
    for content in assistant:
        expanded_rows.append({'document': row['document'],'reference': row['user'], 'assistant': content})


expanded_df = pd.DataFrame(expanded_rows)

expanded_df

## | 구분자로 question, answer 삭제
expanded_df[['question', 'answer']] = expanded_df['assistant'].str.split('|', expand=True)
del expanded_df['assistant']

## 앞에 나열하는 번호 삭제
expanded_df['question'] = expanded_df['question'].str.replace(r'^- ','',regex = True)
expanded_df['question'] = expanded_df['question'].str.replace(r'^(\d\. )','',regex = True)

expanded_df.columns = ['document','reference','question', 'answer']

expanded_df.to_excel('./국방부훈령_질문 증강.xlsx', index = False)
